Convert times tests to mocha
