+++ test/test-async.js
@@ -413,103 +413,6 @@ var console_fn_tests = function(name){
 
 };
 
-
-exports['times'] = {
-
-    'times': function(test) {
-    test.expect(2);
-    async.times(5, function(n, next) {
-        next(null, n);
-    }, function(err, results) {
-        test.ok(err === null, err + " passed instead of 'null'");
-        test.same(results, [0,1,2,3,4]);
-        test.done();
-    });
-},
-
-    'times 3': function(test){
-    test.expect(1);
-    var args = [];
-    async.times(3, function(n, callback){
-        setTimeout(function(){
-            args.push(n);
-            callback();
-        }, n * 25);
-    }, function(err){
-        if (err) throw err;
-        test.same(args, [0,1,2]);
-        test.done();
-    });
-},
-
-    'times 0': function(test){
-    test.expect(1);
-    async.times(0, function(n, callback){
-        test.ok(false, 'iteratee should not be called');
-        callback();
-    }, function(err){
-        if (err) throw err;
-        test.ok(true, 'should call callback');
-    });
-    setTimeout(test.done, 25);
-},
-
-    'times error': function(test){
-    test.expect(1);
-    async.times(3, function(n, callback){
-        callback('error');
-    }, function(err){
-        test.equals(err, 'error');
-    });
-    setTimeout(test.done, 50);
-},
-
-    'timesSeries': function(test){
-    test.expect(2);
-    var call_order = [];
-    async.timesSeries(5, function(n, callback){
-        setTimeout(function(){
-            call_order.push(n);
-            callback(null, n);
-        }, 100 - n * 10);
-    }, function(err, results){
-        test.same(call_order, [0,1,2,3,4]);
-        test.same(results, [0,1,2,3,4]);
-        test.done();
-    });
-},
-
-    'timesSeries error': function(test){
-    test.expect(1);
-    async.timesSeries(5, function(n, callback){
-        callback('error');
-    }, function(err){
-        test.equals(err, 'error');
-    });
-    setTimeout(test.done, 50);
-},
-
-    'timesLimit': function(test){
-    test.expect(7);
-
-    var limit = 2;
-    var running = 0;
-    async.timesLimit(5, limit, function (i, next) {
-        running++;
-        test.ok(running <= limit && running > 0, running);
-        setTimeout(function () {
-            running--;
-            next(null, i * 2);
-        }, (3 - i) * 10);
-    }, function(err, results){
-        test.ok(err === null, err + " passed instead of 'null'");
-        test.same(results, [0, 2, 4, 6, 8]);
-        test.done();
-    });
-}
-
-};
-
 console_fn_tests('log');
 console_fn_tests('dir');
 /*console_fn_tests('info');
